refactor(composites): consolidate architectures, derivers, + register ParCa
Behavior-preserving composite cleanup (baseline byte-identical across a
120 sim-second deep-signature parity gate; 352 tests collect clean).

Architectures
- Remove departitioned + reconciled composites entirely, plus the whole
  cross-architecture comparison cluster (compare viz, compare_report /
  network_report / viz_network, test_architectures_grow,
  test_visualizations_compare, architectures-compare investigation +
  golden fixture + CI list entries). Registry collapses to
  {baseline, colony, millard_pdmp_baseline, parca}.
- Register the ParCa pipeline as a first-class @composite_generator
  ("parca") peer via core_extensions (register_parca_core: step classes
  + schema). build_composite("parca") returns the structural document
  without firing the multi-hour fit; running still goes through
  v2ecoli-parca / build_parca_composite.

Derivers (formerly "listeners")
- Rename steps/listeners -> steps/derivers; Mass*Listener -> Mass*Deriver;
  ListenerStore type -> DerivedProperties (listener_store + ListenerStore
  kept as back-compat aliases). The `listeners` wire path is preserved, so
  emitter/dashboard/parity/existing Parquet keep working.
- Drop redundant per-leaf overwrite[] wrappers where the DerivedProperties
  container already does leaf-replace (mass_deriver, dna_supercoiling);
  keep them where additive apply would accumulate (count/array derivers).
- Consolidate RNACounts + MonomerCounts + UniqueMoleculeCounts into a
  single CountsDeriver (genuine single step, logic inlined verbatim).
  Migrate baseline + millard_pdmp_baseline to counts_deriver.
