+++ .github/workflows/ci.yml
@@ -92,8 +92,6 @@ jobs:
               'v2ecoli/core.py',
               'v2ecoli/composites/baseline.py',
               'v2ecoli/composites/colony.py',
-              'v2ecoli/composites/departitioned.py',
-              'v2ecoli/composites/reconciled.py',
               'v2ecoli/composites/millard_pdmp_baseline.py'
             ) }}
