--- docs/workspace-template/investigations/architectures-compare/spec.yaml
@@ -1,113 +1,0 @@
-# Skeleton Investigation spec for v2ecoli — three-architecture comparison.
-#
-# Runs baseline (partitioned, 55 procs), departitioned (41 steps fused), and
-# reconciled (hybrid) at a common seed/duration, side-by-side. The catch-
-# regressions Study: if departitioned or reconciled drifts from baseline on
-# mass/growth/RNAP, this is where it shows up first.
-#
-# Requires V2ECOLI_CACHE pointing at v2ecoli/out/cache (built once via
-# `python scripts/build_cache.py`). All three architectures consume the same
-# ParCa cache.
-
-name: architectures-compare
-question: |
-  Do departitioned and reconciled match baseline on mass, growth rate, and
-  RNAP count over a full cycle?
-hypothesis: |
-  All three architectures agree within 5 % on cell_mass and growth_rate
-  trajectories and within 10 % on active RNAP count at division.
-status: draft
-tags: [comparison, regression-gate, all-architectures]
-
-baseline: baseline
-
-variants:
-  - name: baseline
-    source: local:baseline
-    parameters: {seed: 0, cache_dir: ${V2ECOLI_CACHE}}
-
-  - name: departitioned
-    source: local:departitioned
-    parameters: {seed: 0, cache_dir: ${V2ECOLI_CACHE}}
-
-  - name: reconciled
-    source: local:reconciled
-    parameters: {seed: 0, cache_dir: ${V2ECOLI_CACHE}}
-
-observables:
-  - {path: [agents, "0", listeners, mass, cell_mass]}
-  - {path: [agents, "0", listeners, mass, dry_mass]}
-  - {path: [agents, "0", listeners, mass, instantaneous_growth_rate]}
-  - {path: [agents, "0", listeners, mass, volume]}
-  - {path: [agents, "0", listeners, mass, protein_mass]}
-  - {path: [agents, "0", listeners, mass, rRna_mass]}
-  - {path: [agents, "0", listeners, rnap_data, n_active_rnap]}
-  - {path: [agents, "0", divide]}
-
-runs:
-  - {composite: baseline,      params: {seed: 0}, steps: 2520}
-  - {composite: departitioned, params: {seed: 0}, steps: 2520}
-  - {composite: reconciled,    params: {seed: 0}, steps: 2520}
-
-comparisons:
-  - name: arch-mass
-    description: "Cell mass over time, all three architectures."
-    variants: [baseline, departitioned, reconciled]
-    observables: [cell_mass]
-  - name: arch-growth
-    description: "Instantaneous growth rate, all three architectures."
-    variants: [baseline, departitioned, reconciled]
-    observables: [instantaneous_growth_rate]
-  - name: arch-rnap
-    description: "Active RNAP count — sensitive growth-rate indicator (see user_preferences memory)."
-    variants: [baseline, departitioned, reconciled]
-    observables: [n_active_rnap]
-
-visualizations:
-  # CompareVisualization takes a list of composite specs and renders the
-  # canonical v2ecoli architecture-comparison HTML (legacy compare_report).
-  - name: compare-report
-    address: "local:CompareVisualization"
-    inputs:
-      composite_specs:
-        source: variant-documents
-        variants: [baseline, departitioned, reconciled]
-    config:
-      title: "v2ecoli — baseline vs departitioned vs reconciled"
-
-  # Per-comparison timeseries traces. TimeSeriesPlot understands the
-  # `comparison:` shorthand (study-model-design.md) and pulls sources +
-  # observable from the named comparison object.
-  - name: mass-cmp
-    address: "local:TimeSeriesPlot"
-    config: {comparison: arch-mass,   title: "Cell mass — three architectures"}
-  - name: growth-cmp
-    address: "local:TimeSeriesPlot"
-    config: {comparison: arch-growth, title: "Growth rate — three architectures"}
-  - name: rnap-cmp
-    address: "local:TimeSeriesPlot"
-    config: {comparison: arch-rnap,   title: "Active RNAP — three architectures"}
-
-  # Per-architecture process topology.
-  - name: net-baseline
-    address: "local:NetworkVisualization"
-    inputs: {composite_spec: {source: variant-document, variant: baseline}}
-    config: {title: "Baseline topology"}
-  - name: net-departitioned
-    address: "local:NetworkVisualization"
-    inputs: {composite_spec: {source: variant-document, variant: departitioned}}
-    config: {title: "Departitioned topology"}
-  - name: net-reconciled
-    address: "local:NetworkVisualization"
-    inputs: {composite_spec: {source: variant-document, variant: reconciled}}
-    config: {title: "Reconciled topology"}
-
-# ── Notes ───────────────────────────────────────────────────────────────────
-# - The RNAP path uses listeners.rnap_data.n_active_rnap rather than the unique
-#   array path; the listener exists in all three architectures and exposes a
-#   scalar count without an aggregator hop.
-# - All three composite generators consume the same ParCa cache, so cache-rebuild
-#   cost is amortised across this Study.
-# - If a viz fails with "no comparison named X", the orchestrator may not yet
-#   resolve `config.comparison:` → sources/observable; fall back to the
-#   explicit-sources form (config: {sources: [baseline, …], observable: …}).
