--- tests/test_composites_departitioned.py
@@ -1,34 +1,0 @@
-"""Unit tests for v2ecoli.composites.departitioned."""
-
-import os
-
-import pytest
-
-
-@pytest.mark.fast
-def test_departitioned_function_is_registered():
-    from pbg_superpowers.composite_generator import _REGISTRY
-    from v2ecoli.composites import departitioned  # noqa: F401
-    names = {e.name for e in _REGISTRY.values()}
-    assert "departitioned" in names
-
-
-@pytest.mark.fast
-def test_departitioned_function_signature():
-    import inspect
-    from v2ecoli.composites.departitioned import departitioned
-    sig = inspect.signature(departitioned)
-    assert set(sig.parameters) == {"core", "seed", "cache_dir"}
-
-
-@pytest.mark.sim
-def test_departitioned_returns_a_document():
-    if not os.path.isdir("out/cache") and not os.environ.get("CI"):
-        pytest.skip("cache dir 'out/cache' not present; "
-                    "build via `python scripts/build_cache.py` (CI builds it automatically)")
-    from v2ecoli.core import build_core
-    from v2ecoli.composites.departitioned import departitioned
-    core = build_core()
-    doc = departitioned(core=core, seed=0, cache_dir="out/cache")
-    assert isinstance(doc, dict)
-    assert len(doc) > 0
