+++ tests/test_seed_determinism.py
@@ -9,8 +9,8 @@ reproducibility claim downstream.
 
 Skipped if `out/cache` is not present, matching test_architectures_grow.py.
 
-Only the baseline architecture is checked here; extend to departitioned /
-reconciled if determinism in those variants also becomes a concern.
+Only the baseline architecture is checked here; extend to other
+architectures if determinism in those variants also becomes a concern.
 """
 from __future__ import annotations
