+++ v2ecoli/composites/millard_pdmp_baseline.py
@@ -143,9 +143,9 @@ BASE_EXECUTION_LAYERS = [
     ['lqr-controller'], FLUSH,
 
     # Layer 7: listeners
-    ['RNA_counts_listener', 'ecoli-mass-listener',
-     'monomer_counts_listener', 'replication_data_listener', 'ribosome_data_listener',
-     'rna_synth_prob_listener', 'rnap_data_listener', 'unique_molecule_counts'], FLUSH,
+    ['counts_deriver', 'ecoli-mass-listener',
+     'replication_data_listener', 'ribosome_data_listener',
+     'rna_synth_prob_listener', 'rnap_data_listener'], FLUSH,
 
     # Emitter + clock
     ['emitter'],
@@ -398,15 +398,12 @@ def _get_step_config(
     from v2ecoli.processes.tf_unbinding import TfUnbinding
     from v2ecoli.processes.chromosome_structure import ChromosomeStructure
     from v2ecoli.steps.partition import Requester, Evolver
-    from v2ecoli.steps.listeners.mass_listener import MassListener, PostDivisionMassListener
-    from v2ecoli.steps.listeners.rna_counts import RNACounts
-    from v2ecoli.steps.listeners.rna_synth_prob import RnaSynthProb
-    from v2ecoli.steps.listeners.monomer_counts import MonomerCounts
-    from v2ecoli.steps.listeners.dna_supercoiling import DnaSupercoiling
-    from v2ecoli.steps.listeners.replication_data import ReplicationData
-    from v2ecoli.steps.listeners.rnap_data import RnapData
-    from v2ecoli.steps.listeners.unique_molecule_counts import UniqueMoleculeCounts
-    from v2ecoli.steps.listeners.ribosome_data import RibosomeData
+    from v2ecoli.steps.derivers.mass_deriver import MassDeriver, PostDivisionMassDeriver
+    from v2ecoli.steps.derivers.rna_synth_prob import RnaSynthProb
+    from v2ecoli.steps.derivers.dna_supercoiling import DnaSupercoiling
+    from v2ecoli.steps.derivers.replication_data import ReplicationData
+    from v2ecoli.steps.derivers.rnap_data import RnapData
+    from v2ecoli.steps.derivers.ribosome_data import RibosomeData
     from v2ecoli.steps.media_update import MediaUpdate
     from v2ecoli.steps.exchange_data import ExchangeData
 
@@ -476,6 +473,24 @@ def _get_step_config(
         out_topo = {"bulk": ("bulk",)}
         return instance, instance.topology, 'step', in_topo, out_topo
 
+    # Consolidated counts deriver: one step computing the RNA / monomer /
+    # unique-molecule count readouts (byte-identical to the three former
+    # listeners). Assemble its config from the three former config names.
+    if step_name == 'counts_deriver':
+        from v2ecoli.steps.derivers.counts_deriver import CountsDeriver
+        merged_cfg = {}
+        for cfg_name in ('RNA_counts_listener', 'monomer_counts_listener',
+                         'unique_molecule_counts'):
+            try:
+                merged_cfg.update(loader.get_config_by_name(cfg_name) or {})
+            except (KeyError, AttributeError):
+                pass
+        instance = _make_instance(CountsDeriver, merged_cfg, core)
+        topology = getattr(instance, 'topology', {})
+        if callable(topology):
+            topology = topology()
+        return instance, topology, 'step'
+
     try:
         config = loader.get_config_by_name(base_name)
     except (KeyError, AttributeError):
@@ -503,15 +518,12 @@ def _get_step_config(
     }
 
     SIMPLE_STEPS = {
-        'ecoli-mass-listener': MassListener,
-        'post-division-mass-listener': PostDivisionMassListener,
-        'RNA_counts_listener': RNACounts,
+        'ecoli-mass-listener': MassDeriver,
+        'post-division-mass-listener': PostDivisionMassDeriver,
         'rna_synth_prob_listener': RnaSynthProb,
-        'monomer_counts_listener': MonomerCounts,
         'dna_supercoiling_listener': DnaSupercoiling,
         'replication_data_listener': ReplicationData,
         'rnap_data_listener': RnapData,
-        'unique_molecule_counts': UniqueMoleculeCounts,
         'ribosome_data_listener': RibosomeData,
         'media_update': MediaUpdate,
         'exchange_data': ExchangeData,
