+++ v2ecoli/library/cache_version.py
@@ -55,8 +55,6 @@ INPUT_FILES: tuple[str, ...] = (
     # old architecture.
     "v2ecoli/composites/baseline.py",
     "v2ecoli/composites/colony.py",
-    "v2ecoli/composites/departitioned.py",
-    "v2ecoli/composites/reconciled.py",
     "v2ecoli/composites/millard_pdmp_baseline.py",
 )
