--- v2ecoli/steps/listeners/rna_counts.py
@@ -1,172 +1,0 @@
-"""
-====================
-RNA Counts Listener
-====================
-"""
-
-import numpy as np
-from v2ecoli.library.schema import numpy_schema, attrs, listener_schema
-from v2ecoli.library.schema_types import RNA_ARRAY
-from v2ecoli.library.ecoli_step import EcoliStep as Step
-
-# topology_registry removed — topology defined as class attribute
-
-
-NAME = "RNA_counts_listener"
-TOPOLOGY = {
-    "listeners": ("listeners",),
-    "RNAs": ("unique", "RNA"),
-    "global_time": ("global_time",),
-    "timestep": ("timestep",),
-}
-
-
-class RNACounts(Step):
-    """
-    Listener for the counts of each mRNA and rRNA transcription units and
-        cistrons. Includes the counts of both partial and full transcripts.
-    """
-
-    name = NAME
-    topology = TOPOLOGY
-
-    config_schema = {
-        'rna_ids': 'list[string]',
-        'mrna_indexes': 'array[integer]',
-        'all_TU_ids': 'list[string]',
-        'mRNA_indexes': 'array[integer]',
-        'mRNA_TU_ids': 'list[string]',
-        'rRNA_indexes': 'array[integer]',
-        'rRNA_TU_ids': 'list[string]',
-        'all_cistron_ids': 'list[string]',
-        'cistron_is_mRNA': 'array[integer]',
-        'mRNA_cistron_ids': 'list[string]',
-        'cistron_is_rRNA': 'array[integer]',
-        'rRNA_cistron_ids': 'list[string]',
-        'cistron_tu_mapping_matrix': 'csr_matrix',
-        'time_step': 'float{1.0}',
-        'emit_unique': 'boolean{false}',
-    }
-
-
-    def inputs(self):
-        return {
-            'RNAs': {'_type': RNA_ARRAY, '_default': []},
-            'global_time': {'_type': 'float', '_default': 0.0},
-            'timestep': {'_type': 'float', '_default': 1.0},
-        }
-
-    def outputs(self):
-        return {
-            'listeners': {
-                'rna_counts': {
-                    'mRNA_counts': {'_type': f'overwrite[array[{self.n_mRNA_TU},integer]]', '_default': []},
-                    'full_mRNA_counts': {'_type': f'overwrite[array[{self.n_mRNA_TU},integer]]', '_default': []},
-                    'partial_mRNA_counts': {'_type': f'overwrite[array[{self.n_mRNA_TU},integer]]', '_default': []},
-                    'mRNA_cistron_counts': {'_type': f'overwrite[array[{self.n_mRNA_cistron},integer]]', '_default': []},
-                    'full_mRNA_cistron_counts': {'_type': f'overwrite[array[{self.n_mRNA_cistron},integer]]', '_default': []},
-                    'partial_mRNA_cistron_counts': {'_type': f'overwrite[array[{self.n_mRNA_cistron},integer]]', '_default': []},
-                    'partial_rRNA_counts': {'_type': f'overwrite[array[{self.n_rRNA_TU},integer]]', '_default': []},
-                    'partial_rRNA_cistron_counts': {'_type': f'overwrite[array[{self.n_rRNA_cistron},integer]]', '_default': []},
-                },
-            },
-        }
-
-
-    def initialize(self, config):
-
-        # Get IDs and indexes of all mRNA and rRNA transcription units
-        self.all_TU_ids = self.parameters["all_TU_ids"]
-        self.mRNA_indexes = self.parameters["mRNA_indexes"]
-        self.mRNA_TU_ids = self.parameters["mRNA_TU_ids"]
-        self.rRNA_indexes = self.parameters["rRNA_indexes"]
-        self.rRNA_TU_ids = self.parameters["rRNA_TU_ids"]
-
-        # Get IDs and indexes of all mRNA and rRNA cistrons
-        self.all_cistron_ids = self.parameters["all_cistron_ids"]
-        self.cistron_is_mRNA = self.parameters["cistron_is_mRNA"]
-        self.mRNA_cistron_ids = self.parameters["mRNA_cistron_ids"]
-        self.cistron_is_rRNA = self.parameters["cistron_is_rRNA"]
-        self.rRNA_cistron_ids = self.parameters["rRNA_cistron_ids"]
-
-        # Get mapping matrix between TUs and cistrons
-        self.cistron_tu_mapping_matrix = self.parameters["cistron_tu_mapping_matrix"]
-
-        # Dimension variables for shaped array types
-        self.n_mRNA_TU = len(self.mRNA_TU_ids)
-        self.n_rRNA_TU = len(self.rRNA_TU_ids)
-        self.n_mRNA_cistron = len(self.mRNA_cistron_ids)
-        self.n_rRNA_cistron = len(self.rRNA_cistron_ids)
-
-    def update_condition(self, timestep, states):
-        return (states["global_time"] % states["timestep"]) == 0
-
-    def update(self, states, interval=None):
-        # Guard: return empty on first tick if data not yet populated
-        # Get attributes of mRNAs
-        TU_indexes, can_translate, is_full_transcript = attrs(
-            states["RNAs"], ["TU_index", "can_translate", "is_full_transcript"]
-        )
-        is_rRNA = np.isin(TU_indexes, self.rRNA_indexes)
-
-        # Get counts of mRNA and rRNA transcription units
-        all_TU_counts = np.bincount(
-            TU_indexes[np.logical_or(can_translate, is_rRNA)],
-            minlength=len(self.all_TU_ids),
-        )
-        mRNA_counts = all_TU_counts[self.mRNA_indexes]
-        full_TU_counts = np.bincount(
-            TU_indexes[np.logical_and(can_translate, is_full_transcript)],
-            minlength=len(self.all_TU_ids),
-        )
-        full_mRNA_counts = full_TU_counts[self.mRNA_indexes]
-        partial_TU_counts = all_TU_counts - full_TU_counts
-        partial_mRNA_counts = mRNA_counts - full_mRNA_counts
-        # All unique rRNAs are partially transcribed
-        partial_rRNA_counts = all_TU_counts[self.rRNA_indexes]
-
-        # Calculate counts of mRNA cistrons from transcription unit counts
-        # TODO (ggsun): Partial RNA cistron counts should take into account
-        # 	the lengths of each RNA transcript.
-        cistron_counts = self.cistron_tu_mapping_matrix.dot(all_TU_counts)
-        mRNA_cistron_counts = cistron_counts[self.cistron_is_mRNA]
-        full_mRNA_cistron_counts = self.cistron_tu_mapping_matrix.dot(full_TU_counts)[
-            self.cistron_is_mRNA
-        ]
-        partial_mRNA_cistron_counts = self.cistron_tu_mapping_matrix.dot(
-            partial_TU_counts
-        )[self.cistron_is_mRNA]
-        partial_rRNA_cistron_counts = cistron_counts[self.cistron_is_rRNA]
-
-        update = {
-            "listeners": {
-                "rna_counts": {
-                    "mRNA_counts": mRNA_counts,
-                    "full_mRNA_counts": full_mRNA_counts,
-                    "partial_mRNA_counts": partial_mRNA_counts,
-                    "partial_rRNA_counts": partial_rRNA_counts,
-                    "mRNA_cistron_counts": mRNA_cistron_counts,
-                    "full_mRNA_cistron_counts": full_mRNA_cistron_counts,
-                    "partial_mRNA_cistron_counts": partial_mRNA_cistron_counts,
-                    "partial_rRNA_cistron_counts": partial_rRNA_cistron_counts,
-                }
-            }
-        }
-        return update
-
-
-def test_rna_counts_listener():
-    from ecoli.experiments.ecoli_master_sim import EcoliSim
-
-    sim = EcoliSim.from_file()
-    sim.max_duration = 2
-    sim.raw_output = False
-    sim.build_ecoli()
-    sim.run()
-    listeners = sim.query()["agents"]["0"]["listeners"]
-    assert isinstance(listeners["rna_counts"]["mRNA_counts"][0], list)
-    assert isinstance(listeners["rna_counts"]["mRNA_counts"][1], list)
-
-
-if __name__ == "__main__":
-    test_rna_counts_listener()
